Emacs wrapper: fix initial flash of wrong color
Normally, Emacs starts rendering a frame before it has
run the code to configure it, resulting in a brief
flash of the wrong color - bad when the intended color
is dark, and everything else is dark, and the frame
briefly flashes white.

This fixes that by forcing the initial background to
be black. For a new Emacs process, it uses X resources,
which are applied from the very beginning. For the
Emacs client, it injects frame parameters, which seems
to apply early enough.

Since the initial black text color would be unusuable
on black background without additional config, these
scripts also set the initial text color to light gray.