git diff independent of git
This is something I've been using for years, since
git diff, especially with --color-words, is often
the best diff available on a box.

Without this script, trying to compare files that
are within a git repo will cause git to show the
diff to their older versions, instead of letting
me compare the files to each other.

The simple trick to fix that is to pass absolute
paths to git diff, but doing that manually is
often annoying so this script gets the absolute
path to each argument and passes that to git diff.

In a polished released tool I would give it proper
argument parsing so that options could go before
the files, but this works well enough for me.