Wrap new tools to use color in pipes by default
In the modern world of command-line tools that use color
well, I often want color by default, even in pipes. For
example, when narrowing the output of `fd` or `rg` in
`fzf`, I find it helpful to preserve the colors, and in
only one or two cases have I wanted to disable colors in
a pipeline.

At least that's the idea I'm going with while I play with
these tools. Eventually the ecosystem will starts to
ossify due to people writing code which uses these tools
with the bad practice of leaving the --color option off
instead of explicitly saying --color=never, and then I
will need to reconsider.