`rm -r` for Git repository directories
1. The `.git` folder contains object files that `rm` asks to
   delete unless we add `-f` or do a prior `chmod +w`, but
2. I don't want to use `rm -rf` unless I really need the `-f`,
   because
3. `rm` with `-f` suppresses certain errors which are wrong
   to suppress for this use-case. Chiefly, if I typo the
   directory, it pretends to succeed instead of telling me
   the very relevant error that I was wrong. Bad design
   conflating "force" with "idempotent"/"redundant==success".

We can neatly solve this by redirecting `rm`'s `stdin` to not
be a terminal, which suppresses the "are you sure you want to
delete it?" questions while still reporting all actual errors.

A secondary safety is to first hit the .git subdirectory, and
only if that succeeds, the whole directory. That way, if we
accidentally target a directory that's not a git repository,
it will error out instead of failing, and the error message
is even self-descriptive, about target/.git not existing.