.emacs: refactor vi-in-minibuffer indicator
Annoyingly repetitive and redundant, but much more
conveniently malleable at runtime, which I think I
like more. (The colors could be pulled out into
variables, which positions this more for reuse by
others as a configurable out-of-the-box package.
The whole indicator logic can be changed without
editing the hooks, by just redefining the function,
which is nicer for when I next fiddle with it.)
Also, this can handle indicating unexpected states.