.emacs: consult-history-execute(-or-not!)
Improved my `search-history-and-run` wrapper
for `consult-history` (also renamed to it to
`consult-history-execute`) to support "q" in
vi command mode in the minibuffer to exit as
`consult-history` normally does, without
executing.

"q" also exits in command mode instead of in
insert mode, in part because it seems better
for versatility of editing the command-line
and because you just went to command mode to
hit "q" in the minibuffer, but also in part
because this provides a really neat option
for repeated commands - RET will execute the
command and go into insert state and then a
repetition will also keep going into insert
mode; "q RET" will execute the command and
stay in command state and then ". RET" will
also keep staying in command state.

In order to make that last bit work, RET in
Eshell (eshell-send-input) has to be marked
with Evil as a non-(repeatable-)command, or
"." would repeat the "RET" (with an empty
line (in all but unrealistic corner-cases)).