.emacs: improve eshell vi state-change automation
Instead of starting over in vi "command" state after
every command, just restore the state from when the
last command was submitted.

Usage quickly reminded that one-off improvised
workflows where each command is not repetitive
are common, and revealed to me how frustrating
and inefficient it can be in those cases to be
taken back to command state each time.