e: use EDITOR environment variable
Conceptually, `e` becomes "editor" shorthand with
an Emacs default rather than Emacs shorthand.

Practically, `e` will open files in my Emacs even
over TRAMP (when using the `with-editor` package).

Big design discomfort was that `with-editor` only
works for remote files if the EDITOR environment
variable gets evaluated by a shell. Git does that.
visudo does not. I might run `e` as a privileged
user. But most privilege-escalation programs will
not let through an unprivileged EDITOR value. And
`exec "$EDITOR"` is just as exploitable as `eval
"exec $EDITOR"` unless EDITOR names a program at
a path that the setter of EDITOR cannot write or
replace. But it is easier and simpler to do your
own PATH lookup and check relevant permissions
than to check for possibly unsafe sh syntax, so
accepting `sh` syntax inherently requires greater
complexity to secure to the same extent. Finally,
security should restrict capability, not ease and
flexibility of expressing allowed actions.

I think the balance of factors leans just enough
in favor of the eval, at least until I have Emacs
code that can drop a with-editor program and add
it to PATH when TRAMPing remotely or privilegedly.