.emacs: fix comint prompt catching vi state color
Turns out comint-highlight-prompt face inherits from
minibuffer-prompt, which seems a bit incorrect to me.

It seems like the intent and correct semantics here
is "whatever the minibuffer prompt was configured to
look like is probably a good default for the look
of a REPL prompt", not "if the minibuffer has state
or events that are best expressed by changing colors
on its prompt, that information should also be shown
on ever REPL prompt by default".

In any case, that's what I think is best, so for now
I'm copying the foreground color attribute before I
start coloring the minibuffer prompt - a more proper
solution might be a full copy of all face attributes
and unsetting the inheritance, but I'll wait and see
if it ever matters one way or the other.