.emacs: fix inconsistent mode line colors
Emacs seems to think that "#000000" is more of an
abstract color feeling - not a precise color code.

Specifically, it seems to look up the basic palette
listed for TERM and see if #000000 is in there. So
it might emit the code for that terminal's palette
"black" instead of #000000, possibly producing bad
colors if the user has customized the palette.

Rather than try to tell Emacs my terminal's palette,
I think the best bet for consistent colors in all
situations is to just make Emacs use almost-black.