.emacs: add motion state to vi/Evil color-coding
Without this, now that the coloring is fixed to not
need `post-command-hook`, switching to motion state
from any other state doesn't switch color.

And since I'm adding a motion state hook, might as
well add a dedicated color for it (the solid white
was originally meant as an error color, I just let
it kick in for motion state during the first pass
of modeline coloring because that was good enough).