.emacs: fix with-editor deleting COMMIT_EDITMSG file on cancel
Apparently this behavior was added back in
https://github.com/magit/magit/issues/2258
to brute force desired behavior on cancel
despite broken wrapper scripts and such.

For me this just caused me to lose commit text that
I carefully compose over several minutes, because I
did ":w" and then ":q" seperately out of safe habit,
for example.

I hate this solution, but it's the first working
solution I was able to come up with: advice around
`with-editor-return` to temporarily add another
advice around the built-in global `delete-file` to
make it a no-op.