.emacs: disable with-editor saving on cancel
When I cancel an edit, I mean *cancel*, not "save my
edit (and indicate an error which might be taken as
a cancel higher up in the call tree)".

This is more personal/subjective/relative than the
last commit (disabling file deletion), but for me,
there is "save" as one operation, and then there is
the orthogonal "yes/confirm" vs "no/cancel" choice
(this maps poorly to vi-style commands: vi has
"save" and "quit", so a "cancel" at best can be
inferred from a quit without a save - incidentally,
until I learned that it didn't work that way, my
intuition as a vi user was that a ":q!" would find
a way to abort the commit).

In particular, the number one thing I don't want is
to say "nevermind, throw away my changes and exit"
the same way that I've been saying it for over a
decade to my editor, in a way that's so habitualized
that it's as reflexive as putting my arm out to stop
a fall - and have it behave in this one special case
as the exact opposite save-and-exit (worse, unless I
enable persistent undo history, the exit immediately
makes this mistake significantly harder to undo).