.emacs: make ace-window always hint
"Always": if there's any other pane to switch to.

A movement habit predicated on extra state kept
in my working memory ("more than two window panes
open?") is not mentally ergonomic for me.

Especially not while my brain has `other-window`
movement habitualized, but even without that, I
want to be able to do for example " o2" without
having to hold back the "2" in the special case
of no other switch candidates.

But that doesn't apply to the case where there
are no other window panes to switch to: in that
case, Ctrl-o/" o" is probably an accident, and
there's no conflict with intentional switching
muscle memory.