.emacs: ace-window: initial dispatch actions
Haven't wanted most of the default dispatch
actions yet, plus there's decent odds of me
tripping over some of them on accident in
frustrating ways.

The only action I actually want is going back to
the previously focused window (called "flip" in
ace-window, I guess because it's an ergonomic way
for flipping between two windows repeatedly), and
I'm moving that to "o" for maximum accessibility
and speed (starting from my current " o" binding).

I've left in "?" because there's good enough odds
that it proves unexpectedly desirable later, it
seems unlikely to trip me up (worst it does is
widen the minibuffer to show the options in it),
and there's a funny quirk where if it's not bound
then you get the error message

    ace-select-window: Wrong type argument: listp, restart

instead of the correct "error" message

    No such candidate: ?, hit ‘C-g’ to quit.

like for any other unbound character.

Added "q" because I'm getting increasingly used to
quitting that way (though, delightfully, Esc works
out-of-the-box along with Ctrl-g). Using `aw--done`
instead of `keyboard-quit` for that binding, even
though the latter is named as if it's private,
since it appears that's what C-g and Esc get bound
do during ace-window dispatch and that results in
a subtly different behavior (no beep/bell/vibrate).