.emacs: " om" in vi for ace-delete-other-windows
Turns out that " o o" was *not* actually chaining two
ace-window calls (due to the space prefix working
during ace-window dispatch) (which seemed doubtful
and I probably should've doubted it harder than I
did, but it wasn't impossible) - instead, the space
was being ignored as an unbound candidate, and the
second "o" was falling through to ace-window's key
binding for deleting other windows.

Luckily now that I've been forced to consciously
choose an alternative, this seems better: "m" is
mnemonic for maximize, and (assuming Qwerty layout)
"m" is accessible from "o" with enough gap to not
accidentally hit, and is nicer than repeating " o".