.emacs: fix " om" for ace-delete-other-windows
For some reason, the default value of aw-dispatch-list
contains `delete-other-windows` but behaves as if you
called `ace-delete-other-windows`, but if you manually
bind either of those yourself, `delete-other-windows`
will get called without ace dispatch features and
`ace-delete-other-windows` will fail to clear/negate
the uncolor-and-dim effect that was already in effect
before applying a second iteration of the same.

So for some reason, to manually achieve the same thing
on a different keybind, I need to call a function that
first does `aw--done` to clear the existing dispatch
and then `ace-delete-other-windows` to start the new
one.