.emacs: improve command-line search/re-run UX
fixed-consult-history and consult-history-execute
now by default don't grab the command line under
point as the initial search - but can optionally
do so if given the universal prefix argument.

Instead, by default they now use whatever is on
the input line at the end of the buffer (where
you would normally type the next command, and
which will usually be empty when this runs), even
if the point is currently on some other line.

Because several times now, I've gotten junk in my
command line search because my cursor happened to
be on some other line (either because I forgot to
move itm or more typically because I just barely
grazed the mouse pad previously).

Conversely, I can remember zero situations where
I've wanted to grab a command from the scrollback.
In fact, I only implemented support for grabbing
the command under point even if that's some line
in the scrollback because consult-history did so.
If that wasn't already the behavior, I probably
wouldn't have thought of it or decided it was a
sufficiently right behavior to implement. But it
does seem just barely likely enough to be useful
to keep now that it's already implemented.