.emacs: try swap and copy ace-window actions
Brainstorming how I might use these to decent
effect with Dired and so on, and putting them
in to play with them more generally. Mnemonic:
"x" looks like a swapping; "c" for copy.

I was holding off on adding swap as a binding
even though it seemed better as " ox" than as
" n o" until I had more reason to add it, and
I now see decent reason to try " oc" for copy
with `dired-dwim-target`, and together they
have enough reason to be added.

Also bound "n" to switch back to the normal
window switch operation, for consistency and
because it makes it faster to undo mistakes
when in ace-window dispatch.