.emacs: fix ace-window Evil state cleanup
Prior to this commit, "q" and Esc were manually
bound to clear `override-evil-mode-line-tag`,
but Ctrl-g bypassed that (and binding Ctrl-g in
`aw-dispatch-list` didn't do anything).

I am also unsure if I've stumbled into other ways
to exit which bypassed the cleanup, or if in all
those times I was just hitting Ctrl-g out of
habit to get out the whatever state I was in.

In any case, this seems to robustly fix that. (I
previously tried the same `unwind-protect` fix
and it didn't seem to do anything, but I guess I
just didn't reload the right parts of my config
at the time.)