.emacs: try out shortcuts for window jumps
Super undecided about these, but committing them
in case I end up wanting them. The idea is not to
save a keystroke vs going through " o", the idea
is to maybe make " o" enter repeating mode by
default, at which point it would be annoying to
hit "q" or Esc to finish every window switch, so
then these shortcuts would compensate for that.