.emacs: [olsScCkK] in single-window window state
This makes all of the above key bindings work in
window state even if there is only one window.

So far I have no need for this, and this might
be overvaluing consistency over simplicity of
implementation, but consistency has a tendency
to enable muscle memory and habits to work as
intended even in the edge cases, and it seemed
like a good exercise for refining what I have
designed and implemented so far.