.emacs: 2nd single-window window state refactor
Instead of coding an imitation of `aw-select` to
be used when there's only one window, just lie
to `aw-select about` the number of windows by
hacking `aw-window-list` to return two of the
same window if it would only return one window.

The only bug is that when there is only one
window, the number hint sometimes says "2"
instead of "1" (both "1" and "2" work to
target the window).

This also fixes the bug introduced in the last
refactor, where killing windows in "K" (window
delete, repeatable) down to just one window
would lose the color-coding and mode indicator
(even though that refactor intentionally added
the behavior of remaining in window state in
that case).