.emacs: fix single-window window state compatibility
For some reason, advice on `length` did not seem to
work in some of my Emacs instances (all 28.1+, and
maybe it's a pre-compiling packaged-config issue? I
don't have time for the deep and broad learning it
would likely need to figure that out).

In general, I hate this strategy of hackily advicing
(spelling intentional: adding Emacs Lisp Advice, not
advising as one would a person) functions to change
the window list as needed to force `aw-select` to do
the desired behavior, in a way that's so coupled to
the implementation, among other flaws. In principle
this should just be one simple flag in ace-window -
but until that happens, this works, and gets me the
polished external UI/UX that I want.