.emacs: bind Esc in undo-tree
Only didn't do it this way earlier because I
didn't know how to achieve what I wanted: to
bind Esc in a way that won't interfere with
escape-containing inputs (like how arrow keys
are encoded in terminals).

Decided to go with change-committing quit
instead of abort because most of the time
that I want to exit undo-tree, I want to
keep the changes, and Esc (especially in
the CapsLock position like I have it) is
the key I'm most used to hitting to exit.
(Also, an argument could be made that
undo-tree is like insert state, where Esc
leaves the state but keeps the changes to
the buffer - this is debatable because we
can also see undo-tree as an unfinished
command which shows a preview of what it
will do, like the ":s" ex-command, but I
think undo-tree is more like insert state
in practice - the main reason is how it
gets used, but also the way it looks more
like swapping in changes rather than
previewing what a change might look like.)