.emacs: improve vi-style minibuffer exiting
Esc now does the same thing as "q": quit
in the way most appropriate for the mode
rather than just abort.

The refactor here reduces how easy it is to
granularly evaluate this part of the config
live, but in this case that's actually a
good thing: these lines only do the right
thing when executed in the minibuffer setup
hook.