.emacs: fix "helpful" scroll behavior
I have never understood the appeal of this
larger scroll behavior - all it does is
reduce control over what is in view when,
which I can tune out on large screens but
on small screens it's infuriating.

I think all the benefits of this can be
had with various vi motions anyway? I
guess we'll see if I start to miss it now
that I've tried it for a month or so.