.emacs: use purple for window swap state
Purple is more distinct from the default gray
than the dimmed/desaturated cyan which I had
for it so far, especially in peripheral vision.

Also, this wasn't enough of a reason in itself,
but for me there is weird intuitive similarity
between the color I use for Evil's Emacs state
and swapping windows, more so than the color I
use for motion state (probably because Evil's
Emacs state is something that toggles with all
of the other Evil states, so in my mind it's
strongly associated with a "swap"-like action).