.emacs: tiny cleanup (debatable)
Now that I've got a top-level `use-package` entry
for Dired, no need to have it inside Evil's entry.
(All my other configs are like this so far: later
entries can just assume the effects of prior
entries. In principle and in the long term that's
probably not the best practice, but for now in a
personal config it's good enough.)