.emacs: fix dired's vi state
Since dired was one of the first things I tried
with Evil during this foray into using Emacs, I
had been under the impression that it started
in command state on purpose. But in retrospect
that was probably just the wrong state for it,
(does Evil even give it a state out-of-the-box?)
as seems to be the case with some other modes.