.emacs: make Ctrl-Space work as space prefix everywhere
I find myself wanting this escape A LOT in any buffer
that's in Emacs state, but I have so far never wanted
the default binding (not since I got vi "marks" and
"visual" (text selecting) state, anyway).

And then when I'm mentally "near" need to use it (for
example I'm in a vi-state buffer, but just came from
or am about to go to an Emacs-state buffer), I end up
hitting it sometimes, and it's mentally less overhead
to maintain one less latch/switch/branch about what
buffer I'm in (Ctrl+Space as the universal catch-all,
so I only need to reliably throw in a Ctrl if needed,
not also reliably *omit* Ctrl if it's not needed).

I don't currently have enough want/need/priority/room to
learn the workflow wisdom/vision of `set-mark-command`.