.emacs: fix over-eager eshell-send-input Enter key
Eshell's Enter key indiscriminately sends the
current line as a command, even if the cursor is
in some previous command line or command's output.

I'm sure there are times when this makes sense,
but I can just use " ngh" or "yyGp" or similar
before Enter in vi command mode if I want that.