.emacs: fix Enter losing last vi state in Eshell
Initially I just made it so that Evil state is
saved after each command is fired off. But in
practice, I don't want Enter with empty input
to reset back to the state before the last
command - if I've changed the state and hit
Enter, that's the state I want it in after the
Enter.

The reason for still keeping the pre-command
hook is so that this still works as expected
even if I ever end up starting commands some
other way.