.emacs: reorder vi minibuffer keybinding config
Put it towards the end because it ends up bringing
together multiple unrelated (except by connection
to Evil) pieces of config.

The other (probably more ideal in principle) way
would be to put the minibuffer setup hook at the
start, but enhance it to bind pluggable dispatch
functions for the keys - so for example "j"
would be bound to a new function named something
like `evil-minibuffer-next-line`, and later in
the config I would call a new macro to hook
`'evil-vertico-next-line` into that function with
the predicate `t`; Esc would be bound to another
named `evil-minibuffer-quit` and a macro call
would set `consult-history-execute-quit` with
`(eq this-command 'consult-history-execute)` as
the predicate and `'abort-minibuffers` with `t`
as the predicate.

But I'll probably put that off until my need
starts to exceed the current hard-coded setup.