.emacs: fix vi Esc in Eshell while command runs
My prior work: when Eshell runs a command, the
current Evil state is remembered, Evil is put
into insert state until the command finishes,
and then the remembered state is restored.

Problem with prior work: hitting Esc while a
command is running to go back to Evil command
state does not override the remembered state,
so if I started the command in insert state
and then decided while it runs that I want to
be in command state, I get yanked back to
insert state when the command finishes.

This commit fixes that.