.emacs: workarounds for Android clipboard
Problem 1:

On Android, copy+paste is pretty paternalistically
broken - programs must have input focus to read
the clipboard, which would be the right default,
but instead it's the only option and there is no
way for the user to just tell Android that a given
app is trusted to read the clipboard.

The closest thing to a solution includes using ADB
to turn on a hidden permission to read the logs,
and then have the app sniff the logs for clipboard
logs which include the entire pasticle. Even if
Termux could be made to use that workaround, the
one time I tried a clipboard manager with it, the
strangest bug happened: the clipboard manager was
only catching my pasticles on its own from the
background every other screen unlock.

So the closest workaround to this that I have is
an utter hack: read the clipboard in a floating
window Termux in a loop, minimizing the manual
overhead of giving Termux access to each clip:
just touch the Termux:Float circle and make an
immediate tiny swipe. (And if I switch to Termux
for other reasons, even the non-floating one, it
gets clipbpard access automatically.)

Problem 2:

Emacs in the terminal doesn't know how to read
from the system clipboard. (It does know how to
write to the system clipboard.)

A simple standalone solution is to just install
the `xclip` Emacs package (which even knows how
to use Termux' CLI API for clipboard get and set).

But for some reason, this is profoundly slow on
Termux, slowing all pastes to an unusable point.

Specifically, the slowest part is redirecting an
external command's output to an Emacs buffer. So
simply running the same command but redirecting
it to a file, and then reading the file into an
Emacs buffer, is much faster. But still too slow
for the level of responsiveness I need from
paste, especially since with vi-style editing, I
use "paste" for some repetive single-character
thngs.

So working around both of these converges onto
the solution in this commit.