.emacs: fix fixed-consult-history using prior line
For example, run `git log` in a small-enough Eshell,
and scroll until the first part of a line which is
long enough to wrap is at the very bottom. Press "q"
to exit `git log`, which erases the ":" line (the
prompt from `less`) and replaces it with the Eshell
prompt.

Emacs will remember those last two lines as wrapped,
even though the bottom one is now the shell prompt,
and so certain functions like `beginning-of-line`
will in some cases reach past the prompt and grab
the entire previous line of output. That in turn
causes my `fixed-consult-history` to prepend the
whole prior line to the command selected by
`consult-history`.

(This might be specific to commands handled by Eat
inside Eshell, but I haven't tried to verify that.)