eshell/alias: try vi alias just like e
Emacs is my vi now. So far we're about 4-to-0
on "I typed vi because I wanted to open a file
for editing (in my vi-style Emacs)" vs "I
typed vi because I wanted to open a file in an
external vi specifically".

I held off on this back when I first added the
e alias, but my muscle memory for editing with
"vi" is too strong.