.emacs: make consult-line order more absolute
Determinism which is relativistic to where I am
in the file is less intuitive and ergonomic for
me than something that works the same no matter
my "reference frame".

Concretely: consult-line normally has candidate
lines which are above point sort after all the
candidate lines which are below point. To me
this was initially surprising, and so far when
it makes a difference to me it has always been
a less-helpful difference.