.emacs: Eshell aliases outlived their usefulness
I always suspected this day would come. Eshell
aliases have several signs of being a feature that
was only designed and built to the point of being
good enough for the most super common use-cases.

The deal-breaker is how Eshell aliases (fail to)
combine with `eshell/sudo` - `sudo vi ...` fails
to actually escalate through TRAMP's sudo method
when `vi` is an Eshell alias, but escalates just
fine when `vi` is a function named `eshell/vi`.