.emacs: fix my Esc in operator vi state in Eshell
When I bound Esc to interrupt the current process
in Eshell when in vi command state, I neglected
to "shield" operator state from that rebind
(operator state normally inherits the Esc binding
from command state, so if Esc is rebound in
command state, it should also be explicitly bound
in operator state to its original function of
returning to command state).

Fun story: I discovered this because I had an
`apt-get upgrade` running in Eshell, and I had
put Eshell in command state (to switch to some
other buffer while it ran) - `apt-get` asks me
for a "y/n/..." input, I switch back, hit "y",
I'm not in insert state so the "y" puts me in
operator state, I realize this and try to back
out with Esc, which interrupts the `apt-get`.