.emacs: make Esc to exit i-search
I never really found i-search's lack of a
simple "I'm done, exit i-search" keybind
intuitive or compatible with how I think.

(I don't always want to move my cursor or
do a non-i-search command when I want to
exit i-search, and I rarely want the Ctrl-g
behavior of going back to wherever I was
when I started i-search - though I do like
having that more severe abort option and I
like it on Ctrl-g where it's a little less
fast/easy/smooth than Esc).

Annoyingly, out-of-the-box, Esc *almost*
works to exit i-search (it "works" because
it falls through to whatever it's bound to
under i-search, whether that's the stock
prefix map or Evil's binding of it), so I
already grew a habit for hitting Esc to
drop out of i-search, which I discovered
when Esc in i-search fell through to my
new keybinding in vi motion state to quit
the whole window.