.emacs: make vi replace state Esc to insert state
A very common nuisance I experience with replace
state is reaching the end of text that I want to
overwrite but still having more to type - longer
replacement text than replacement text.

It's fairly ergonomic to tap Esc an extra time
if finishing inserting is not needed after the
replacing - certainly more ergonomic than an Esc
followed by "a" if a finishing insert is needed.

Also, in many cases where I've wanted an insert
after a replace, it would've also been nice for
"." to repeat both as one action. Conveniently,
the simplest implementation gets this behavior
for free.

(Note that this is like	the behavior of replace
state upon reaching the end of a line - this
enables choosing this behavior before reaching
the end of a line as well.)