.emacs: fix vi's cursor backpedaling habit
Not sure if I ever liked this behavior in vi
more than I get inconvenienced by it. It's a
tiny detail either way - time to find out if
I rely on this more than I realize (I suspect
it's negligible either way, but that it is
worse on net and I've just gotten used to
constantly compensating for it so now I will
feel very subtle relief). I do have conscious
memories of some situations where this was a
mild inconvenience.

In any case, I finally got enough reason to
look into it because the last change making
replace state exit into insert state, and in
particular actually using that change, made
me feel the nuisance of cursor backpedaling
on state exit too much in quick succession.
So instead of solving it just for that case,
I'd rather try turning this off globally.