.emacs: stop encoding spaces as tabs
Encoding indentation and tabulation as tab
characters in a raw plaintext file on disk
is not an inherently terrible move, and in
an elastic tabstop world it might even be
good, but my experience is that raw tabs
cause many little problems in practice and
solve approximately zero problems.

Anyway I have been putting up with this
ever since beginning this Emacs "trial"
because I was too busy setting up other
stuff that matters more, but it has once
again been nothing but annoying, and it
breaking my block replacing-paste was
enough to get me to finally take the time
to turn it off.