.emacs: tiny organization tweak
Rapidly approaching the point where I will
need to start separating the Evil section
into sub-sections and adding comments to
remind myself what/why the grouping is.

In the meantime though, an idea I'm having
is that annotating Emacs functions to tell
Evil how to handle their repetition is a
lot like telling Evil what vi state to give
to the various built-in modes, and it makes
sense to put it in a section of its own at
the bottom when it's not part of another
more focused package/feature setup.