.emacs: refactor save-advice -> with-advice
Allow multiple advices in one macro call,
reducing nesting and helping readability.

Same implementation approach as my unpack macros.

Naming is better. I initially wanted to use
"with" but it seemed like it was not a how
Emacs conventions named things - but I now
see that this is a normal naming pattern in
Emacs, and I just misinterpreted; mainly, I
want to note that "save" makes sense in the
`save-excursion` and `save-mutation` macros
because any change of the type they guard is
rolled back - so `save-advice` is appropriate
for a macro which rolls back all advice
changes (whether in general or at least for a
specific function), rather than just rolling
back one advice change.