.emacs: macro refactor and optimization
`let-unpack`, `let-uncons`, and `with-advice`
share the vast majority of their code; I
factored that out into an `apply-split-nest`
macro that applies a list of arguments to
a callable by splitting them into multiple
nested calls to the callable.

In the process I spotted a couple ways to
optimize it down to fewer and lower-level
list operations. Perhaps went too far with
this: now the argument list to be applied
is mutated by the macro, although I am not
sure how to actually create a situation in
which this is a problem - at first naive
glance, `apply-split-nest` won't expand a
variable or evaluate a call that is passed
in as `arguments`, so such invocations will
cleanly error out with "not a list" on any
argument list with life outside this macro.