.emacs: better "G" behavior in vi
This makes "G" behave the way that it normally
works in vi - if the buffer ends in a linefeed
character, it goes into the line prior to that,
instead of the empty space after it.

If nothing else, I use "G" often in help
buffers to get to the "[back]" link, and
overshooting the last line is annoying.

But also just in general this is how I've had
years of expectation for "G" behaving, and I
doubt I'll have many situations where having
"G" overshoot the last substantial line helps.