.emacs: extract reusable comint exit hook
Had this refactor sitting in my Git stash
for a couple weeks or more by now; at the
time I was undecided about it being worth
it and decided to wait for a need; now I
think it's a fine idea. More importantly,
having this functionality buried in the
middle of Evil's config section makes it
very easy to miss or even forget about.