.emacs: change which s-expression cursor "touches"
The native Emacs behavior is that functions like
`eval-last-sexp` will scan for the S-expression
entirely backwards from point.

Evil adjusts this in some modes, because vi's
interface semantics are that the cursor is on
characters rather than between characters; for
example, in normal state `eval-last-sexp` will
scan for the S-expression starting from the
character "under" point (which looks like
"after" point if you have a bar-style cursor).

I find the Evil behavior preferable globally.
In fact I'd say it's objectively more useful:
every S-expression is at least one character
wide, so if you have point "on"/"before" the
last character of an S-expression, the only
useful (not needlessly erroring) behavior is
to include that character in the scan. This
also aligns it with how I set up the cursor
behavior (no backpedaling when exiting insert
state and so on), is more visually obvious in
a terminal where the cursor looks the same in
all modes and thus "touches" the parentheses
in a way that is consistent with normal state.

Implemented with an eye towards reusing it for
other functions and modes. I also initially
went overkill with the special-case handling
for the cursor movement:

    (let ((point-after-adjustment nil))
        (unwind-protect
            (progn
                (condition-case _error
                    (progn
                        (forward-char)
                        (setq point-after-adjustment (point)))
                    (end-of-buffer nil))
                (apply last-s-expression-function arguments))
            (if (equal (point) point-after-adjustment)
                (condition-case _error
                    (backward-char)
                    (beginning-of-buffer nil)))))

This variant would only restore point if
the function which looks for the last
S-expression didn't end with the point
moved from where it started; but ideally
this function would only advise functions
which don't move point, so I decided to go
back to the simpler version in this commit.