.emacs: try evil-move-beyond-eol
This allows the cursor to go beyond the last
character in a line in vi normal-state.

However, "$" and End still work as before.
(Note the parallel with my 'fixed' "G" and
the end-of-file when the file ends in "\n".)
I may gradually force other keys to work
how they work without this setting.

One big reason why I held off on committing
this change despite already trying to try it
a couple times is because I was immediately
dissatisfied with the interactions between
this setting and `eval-last-sexp`. But
that's now fixed as of the last commit.

One caveat with this is that "x" no longer
rule-combos itself into behavior like "X"
at the end of lines. I am fine with this;
if vi was designed from the ground up in a
way that let "x" keep deleting forward at
the line-boundary, I would've never thought
to want the current emergent quirk of how
it behaves in vi/vim at the end of a line.

I can think of a couple of other habits of
mine that this effects, and I'm sure I'll
trip over others soon, but that's fine. I
think there is really big potential that
this is The Right Way, and that my other
habits are actually rather low-value or
compensating for limitations that I can
solve better in other ways in Emacs.