.emacs: try " vd" and " vs" for git (staged) diff
This was an experiment from about a week ago which
I think I will end up discarding. It's not bad but
`git diff-index` is incapable of showing _only_
unstaged changes (without the staged changes), and
I think there is probably something better and
more general for my needs to be achieved by binding
the same keystrokes to a function that creates an
Eat buffer running the desired git commands.